Turn the topics navigation (bottom footer) into an editable MediaWiki: message
The syntax of [[MediaWiki:Webplatform-bottom-footer]] is similar to that
of [[MediaWiki:Sidebar]]:
* Page name|Displayed text
* Another page|some-interface-message-name

The skin class parseItem(), getMessageAsArray() and getBottomFooterLinks()
methods originate from Wikia codebase and thus are licensed under GPLv2+.
See
https://github.com/Wikia/app/blob/dev/includes/wikia/GlobalFunctions.php
for the current Wikia versions of these methods (well, the first two,
anyway) as well as a full history of authors.
These particular methods are copied from ShoutWiki's version of the Monaco
skin and changed to non-static methods (to make use of RequestContext
where appropriate) and old (long) array syntax converted to short array
syntax.

Change-Id: Iad79bd61f679d38ce9eb2b20023f3fd86835bff4