Restore legacy styles removed in upcoming MediaWiki version
Previously certain styles were silently added to all skins. They may
not be needed by your skin, but if they are this is how to restore
them.

These styles however do add 3.75kb to your skin and contain many CSS rules
that are not necessary for modern browsers nor for skins that do not support
languages with RTL content. This is a good opportunity to check in if the
module is actually needed.

More information on the change is given in ResourceLoaderSkinModule and
T242177.

Change-Id: I64b0d57313e929f9f2a718aeda8ec82fb08fe63c