Improve VisualEditor support
VE toolbar now docks to the proper place and "Edit" (w/ VE) as opposed to "Edit source" is the primary "Edit" link in the skin.
That said, the in-place loading still doesn't work with this skin the way it does with MonoBook & friends, but at least VE is usable and works better than prior to this commit.

Change-Id: I6a325e6f3903fe225327b5bba1c28ff40dcadba0