Integrate the BreadcrumbMenu extension by Doug Schepers into the skin
It was always a skin-specific thing, so I didn't see much value in keeping it a separate thing, because building a cross-skin-compatible breadcrumb navigation extension is pretty much impossible due to how vastly different skins in general are.

Furthermore, the skin's description message claims that this is a "skin featuring [--] built-in breadcrumbs" so isn't it only sensible to do what the description message has claimed the skin does since...forever? :-)

The brand new configuration global $wgWebPlatformEnableBreadcrumbs can be used to disable the actual breadcrumb rendering, but the root breadcrumb (which defaults to $wgSitename) will always be rendered, just like previously.

Also added a FIXME note regarding the breadcrumb-related z-index rules, which conflict with OOUI and this is visible on pages like Special:ListUsers where the breadcrumbs somehow, all of a sudden, *overlap* the OOUI menu, which seems awkward. That said, I think this'll require further research, hence why I didn't flat-out remove the rules altogether.

Added myself to the skin authors, since I've made over 40 commits to this repository since 2017, which makes me the developer with the most commits to this repo (if we exclude TWN's automated i18n updates).

Finally bumped up the version number to 1.3, because I think making the top navigation editable and bringing back the breadcrumbs are both significant enough changes to warrant the bump.

Change-Id: Ib1131ef0ff543c2ba2b458f7772381426f6a7341