Make the Orange Bar of Death show up if the user has new messages, even if Echo is installed
By default Echo seems to hide the Orange Bar of Death (which isn't even that orange in this skin).
That works for skins like MonoBook etc. which prominently display the Echo UI icons in the UI.

WebPlatform doesn't do so, because the vast majority of personal tools, including the Echo icons, are hidden in the user drop-down menu.
That is, of course, not ideal. This change allows people to be aware that they have new messages.

An even proper fix would be to move the Echo UI icons out of the aforementioned drop-down menu, but that's significantly trickier to implement.

Change-Id: Icb04ec90aa6d15737486ceeac8c78a2578c98ed5