MediaWiki-fy the footer area for good to make the skin fully reusable
This gets rid of the WebPlatform.org wordmark in the footer area and replaces the listing of steward
organizations with the usual MediaWiki footer links; by default, these are "Privacy policy", "About {{SITENAME}}"
and "Disclaimers".
Additional footer links added by extensions (such as MobileFrontend) are also supported.

Furthermore thisp atch adds the "powered by MediaWiki" button to the footer, right next to the copyright icon (if any),
and a bit of spacing between them.

The ul.stewards element got renamed to ul.footer-links to better describe its current role.

I'm not madly in love with the rule "margin: -2.5em auto 0;" on footer ul.footer-links, but it'll have to do for now.
My guess is that it'll look better if you have ~6 footer links or so instead of the default, aforementioned 3.
Additionally I think that element will need a max-width of some sort defined again, but this current implementation
looks fine on my 720p screen. Your mileage may vary.

Bug: T330251
Change-Id: Iad8956080275494643fe3efe463cc6e38e3e3ea5