Move Echo icons out of the personal tools drop-down menu and next to the user name
It's not perfect, as noted in the LESS file, but it's an improvement, since now the Echo icons are immediately visible on the UI.
This is consistent with MonoBook and other sensible skins, for you should *not* need to hover your mouse over or tap on the user name menu to see if you have new messages/notifications.

The LESS code and the recolored icons are sporked from the Theme extension.

Change-Id: I67c7ae613d03273f8053be87fb291eae2cba7705